Fixed client crash
